Move JSearchHelper to namespace
