+++ components/com_search/views/search/view.html.php
@@ -9,7 +9,6 @@
 
 defined('_JEXEC') or die;
 
-use Joomla\Registry\Registry;
 use Joomla\String\StringHelper;
 
 /**
@@ -106,7 +105,7 @@ class SearchViewSearch extends JViewLegacy
 		$lists['searchphrase'] = JHtml::_('select.radiolist', $searchphrases, 'searchphrase', '', 'value', 'text', $state->get('match'));
 
 		// Log the search
-		JSearchHelper::logSearch($searchWord, 'com_search');
+		\Joomla\CMS\Helper\SearchHelper::logSearch($searchWord, 'com_search');
 
 		// Limit search-word
 		$lang        = JFactory::getLanguage();
